+++ {{cookiecutter.repository_name}}/src/{{cookiecutter.module_name}}/__init__.py
@@ -1,3 +1,0 @@
-"""Top-level module for {{ cookiecutter.module_name }}."""
-
-__version__ = "0.1.0.dev0"