+++ CHANGELOG.md
@@ -81,7 +81,7 @@ The list is not exhaustive as many changes have been made.
 ### Added
 
 - GitHub actions that rely on [City Modelling Lab reusable actions](https://github.com/arup-group/actions-city-modelling-lab).
-- Schema for Cookiecutter input argument config (in `cookiecutter.json`) + tests added to CI to validate the config and the schema itself.
+- Schema for Cookiecutter input argument config (in `json`) + tests added to CI to validate the config and the schema itself.
 - Pre-commit hooks and pre-commit CI
 - Templates for different types of GitHub issues.
 - Cookiecutter contributing guidelines and changelog.